+++ app/src/main/java/chat21/android/demo/usage/SendMessageUsage.java
@@ -2,10 +2,10 @@ package chat21.android.demo.usage;
 
 import android.util.Log;
 
-import chat21.android.core.ChatManager;
-import chat21.android.core.exception.ChatRuntimeException;
-import chat21.android.core.messages.listeners.SendMessageListener;
-import chat21.android.core.messages.models.Message;
+import org.chat21.android.core.ChatManager;
+import org.chat21.android.core.exception.ChatRuntimeException;
+import org.chat21.android.core.messages.listeners.SendMessageListener;
+import org.chat21.android.core.messages.models.Message;
 
 
 /**